requirements.txt
asgiref==3.2.10
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
Click==7.0
colorama==0.4.3
cryptography==2.8
distro==1.4.0
Django==3.1
django-phonenumber-field==6.0.0
django-shopping-cart==0.1
entrypoints==0.3
httplib2==0.14.0
idna==2.8
install==1.3.5
jeepney==0.7.1
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pbr==5.8.0
pexpect==4.6.0
phonenumberslite==8.12.39
Pillow==8.1.0
protobuf==3.6.1
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.21
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-dateutil==2.7.3
python-decouple==3.5
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
testresources==2.0.1
urllib3==1.25.8
wadllib==1.3.3

django-filter~=21.1